Bahri spell4 spell implementation.

+++ Assets/Scripts/ScriptableObjects/SpellData/Bahri/BahriSpell4Data.cs
@@ -10,7 +10,7 @@ public class BahriSpell4Data : SpellData
     [field: SerializeField] public float maxMagnitude { get; private set; }
     [field: SerializeField] public float speed { get; private set; }
     [field: SerializeField] public float radius { get; private set; }
-    [field: SerializeField] public float charges { get; private set; }
+    [field: SerializeField] public int charges { get; private set; }
     [field: SerializeField] public float duration { get; private set; }
     [field: SerializeField] public float takedownDuration { get; private set; }
 }
